% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.streg.R
\name{predict.streg}
\alias{predict.streg}
\title{Predicted Values for a 'streg' Object}
\usage{
\method{predict}{streg}(
  object,
  newdata,
  type = c("lp", "linear", "xb", "hr", "log.p", "p", "log.hazard", "hazard",
    "log.cum.hazard", "cum.hazard", "survival", "quantile", "uquantile", "terms"),
  nocons = FALSE,
  nooffset = FALSE,
  se.fit = FALSE,
  ci.fit = FALSE,
  ci.level = 0.95,
  terms = NULL,
  perc = c(0.1, 0.5, 0.9),
  na.action = na.pass,
  ...
)
}
\arguments{
\item{object}{result of a model fit using the \code{streg} function}

\item{newdata}{data for prediction. If absent predictions are for observations
used in the original fit.}

\item{type}{the type of predicted value. \code{"lp"}, \code{"linear"}, and
\code{"xb"} provide the linear predictor, and \code{"hr"} provides the antilog of
the linear predictor. \code{"p"} and \code{"log.p"} provide the Weibull shape
parameter and it's logarithm, respectively (constant for models that do not include a
\code{\link[survival:strata]{survival::strata()}} term). \code{"log.hazard"} and \code{"hazard"} provide the
hazard on the log and natural scale respectively, as do \code{"log.cum.hazard"} and
\code{"cum.hazard"} for the cumulative hazard. \code{"survival"} provides the survival
probability. \code{"uquantile"} provides the predicted quantile on the linear
predictor scale, and \code{"quantile"} on the data scale. \code{terms} provides the
matrix of centered terms for the linear predictor.}

\item{nocons}{if TRUE removes the intercept from the linear predictor. Only valid for
types \code{"lp"}, \code{"linear"}, \code{"xb"}, and \code{"hr"}.}

\item{nooffset}{if TRUE removes the offset from the linear predictor. Only valid for
types \code{"lp"}, \code{"linear"}, \code{"xb"}, and \code{"hr"}.}

\item{se.fit}{if TRUE, includes the standard errors of the prediction in the result.
The delta method is used for predictions that are non-linear transformations of the
linear predictors.}

\item{ci.fit}{if TRUE, includes a matrix with lower and upper limits for the confidence
interval. For type \code{"hr"}, intervals are calculated on the linear predictor scale
and transformed. Similarly for types \code{"hazard"} and \code{"cum.hazard"}, intervals
are calculated on the log scale and then transformed. Intervals for type
\code{"survival"} are calculated for the log cumulative hazard and transformed.}

\item{ci.level}{level for calculated confidence intervals}

\item{terms}{subset of terms. The default for residual type "terms" is a matrix
with one column for every term, excluding the intercept, in the model.}

\item{perc}{vector of percentiles. This is used only for quantile predictions.}

\item{na.action}{applies only when the newdata argument is present, and defines the
missing value action for the new data. The default is to include all observations.}

\item{...}{for future methods}
}
\value{
A vector or matrix of predicted values. If \code{se.fit} or \code{ci.fit}
are TRUE, the return type is a list containing, where relevant, \code{fit},
\code{se.fit}, and \code{ci.fit} objects.
}
\description{
Predicted values for a \code{streg} object
}
\seealso{
\code{\link[=streg]{streg()}}
}
\author{
David C Muller
}
