\name{rstreg-package}
\alias{rstreg-package}
\alias{rstreg}
\docType{package}
\title{
\packageTitle{rstreg}
}
\description{
\packageDescription{rstreg}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{rstreg}

Maintainer: \packageMaintainer{rstreg}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = rstreg)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{rstreg}}
%\section{Documentation Index}{\packageIndices{rstreg}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
