% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streg.fit.R
\name{streg.fit}
\alias{streg.fit}
\title{Fit a Parametric Survival Model}
\usage{
streg.fit(
  x,
  z,
  y,
  weights,
  offset,
  init,
  pfixed,
  control,
  dist,
  max.method,
  ...
)
}
\arguments{
\item{x}{\code{model.matrix} for the regression parameters.}

\item{z}{\code{model.matrix} for the auxiliary parameters}

\item{y}{\code{Survival outcome of class [survival::Surv()]}}

\item{weights}{observation weights}

\item{offset}{offset for the linear predictor of the regression equation}

\item{init}{vector of initial values}

\item{pfixed}{value to fix auxiliary parameter p at.}

\item{control}{control list passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}}

\item{dist}{survival distribution}

\item{max.method}{maximisation method passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}}

\item{...}{for future methods}
}
\description{
Raw interface for fitting parametric survival regression models. See
\code{\link[=streg]{streg()}} for the formula and data based interface.
}
