% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streg.R
\name{streg}
\alias{streg}
\title{Regression for a Parametric Survival Model}
\usage{
streg(
  formula,
  data,
  weights,
  subset,
  na.action,
  dist = "weibull",
  pfixed = NULL,
  init = NULL,
  max.method = "NR",
  control = NULL,
  model = TRUE,
  x = TRUE,
  z = TRUE,
  y = TRUE,
  robust = FALSE,
  cluster,
  ...
)
}
\arguments{
\item{formula}{a formula expression. The response should be a survival object as returned by the
\code{\link[survival:Surv]{survival::Surv()}} function.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula},
\code{weights}, or \code{subset} arguments.}

\item{weights}{optional vector of observation weights}

\item{subset}{subset of the observations to be used in the fit}

\item{na.action}{a missing-data filter function, applied to the model.frame, after any subset
argument has been used.}

\item{dist}{Parametric survival distribution. Currently only \code{"weibull"} is available.}

\item{pfixed}{optional real value to fix the Weibull parameter p. e.g. set \code{pfixed=1} to
fit an exponential survival model.}

\item{init}{optional vector of initial values for the parameters.}

\item{max.method}{optimisation method passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}}

\item{control}{list of control values passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}}

\item{model}{if TRUE returns the model frame}

\item{x}{if TRUE returns the x matrix}

\item{z}{if TRUE returns the z matrix (predictor matrix for the Weibull parameter p)}

\item{y}{if TRUE returns the response}

\item{robust}{Use robust sandwich variance-covariance instead of the asymptotic
variance-covariance. If there is a cluster argument this defaults to TRUE, and the returned
robust variance-covariance matrix will be adjusted for the number of clusters rather than
the number of observations.}

\item{cluster}{optional variable that identifies groups of observations that is used in the
calculation of the robust variance-covariance}

\item{...}{other arguments passed to \code{\link[maxLik:maxLik]{maxLik::maxLik()}}}
}
\value{
an object of class \code{streg} is returned.
}
\description{
Fit a parametric survival regression model. Currently this function only implements
the Weibull proportional hazards model with parameters \eqn{\lambda} and \eqn{\text{p}}.
This parameterisation is the same as used in Stata's \code{streg} command, but
differs from the parameterisation in \code{\link[survival:survreg]{survival::survreg()}} which uses the location-scale
framework fits models with the accelerated failure time metric. Unlike \code{\link[survival:survreg]{survival::survreg()}},
this function also allows delayed entry/counting process survival data. In future
this function will offer further parametric survival functions in both the proportional
hazards and accelerated failure time metrics.
}
\details{
Currently fits proportional hazards model of the form \eqn{h(\mathbf{t})=h_0(\mathbf{t})\exp(g(\mathbf{X}))}.
The Weibull proportional hazards model has Survivor function \eqn{\exp(-\lambda t^p)}, and
is parameterised by \eqn{\lambda = \exp{\mathbf{X\beta}}}. When \code{formula} contains a
\code{\link[survival:strata]{survival::strata()}} term, the Weibull parameter \eqn{p} is parameterised by
\eqn{p=\exp(\mathbf{Z\gamma})}, thus fitting separate baseline hazards for each level of
the \code{strata} term.
}
\author{
David C Muller
}
